% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Fars Map States}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a state number from 1 to 50 representing one of the states of the United States.}

\item{year}{a year in the format of "yyyy" (multiple years cannot be supplied to this funtion)}
}
\value{
This function returns a plot of the state indicated with all fatalities plotted by latitude/longitude.
An error will be return if an invalid state number is supplied, or the year supplied is not found in the
}
\description{
This functions takes a year and state number as inputs and returns a plot of the fatalities
plotted on a graphic of the state.
}
\examples{
\dontrun{
fars_map_state(25,2013)
}

}
