% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Fars Read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string of the name of the CSV file that you'd like to
read into a dataframe}
}
\value{
This function returns the specified file CSV file as a tbl_df/dataframe object. An
error will be returned if the filename is unable to be found in the working directory.
}
\description{
This is a simple function that reads in a CSV file and returns a tbl_df/dataframe
as it's output.
}
\examples{
\dontrun{
fars_read("example.csv")
}
}
