% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Fars Read Years}
\usage{
fars_read_years(years)
}
\value{
This function returns a list of the year/month found in the FAR file based on the years provided. It will
return an error if the year provided cannot be converted into an integer format or if the filename resulting from
the years specified cannot be found in the directory.
}
\description{
This functions takes years as inputs and returns a list containing the year/month combinations
found in the FAR files
}
\details{
The functions uses the Fars_Read/Make_Filename functions to take the year provided and create a
list of the months/year combinations if found in the FAR files.
}
\examples{
\dontrun{
fars_read_years(2016)
fars_read_years(c(2014,2013))
}
}
