% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Fars Summarize Years}
\usage{
fars_summarize_years(years)
}
\value{
This function returns a tbl_dataframe object totaling the number of fatalities per month for each year specified.
An error will be returned if a year provided cannot be converted to integer format, or if no file corresponding to the
year specified is found in the working directory.
}
\description{
This functions takes years as inputs and returns a tbl_df/dataframe object totaling the number of fatalities
per month for each year specified.
}
\details{
The functions uses the fars_read, make_filename, fars_read_years functions.
This function requires the following external package functions:
}
\examples{
\dontrun{
fars_summarize_years(2016)
fars_summarize_years(c(2014,2013))
}
}
