% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make Filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{a year in the format of "yyyy"}
}
\value{
This function returns the specified filename as a string to be used with the entered year.
This function returns an error if value entered cannot be converted properly into an
integer format.
}
\description{
This function converts a year into the properly formatted filename
}
\examples{
\dontrun{
make_filename(2016)
}
}
